requirements.txt
anyio==3.7.1
astroid==2.15.6
attrs==23.1.0
bcrypt==4.0.1
black==23.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
cssselect==1.2.0
cssutils==2.7.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
exceptiongroup==1.1.2
fabric==3.2.1
fake==0.8
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
invoke==2.2.0
isort==5.12.0
lazy-object-proxy==1.9.0
lxml==4.9.3
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
paramiko==3.3.1
pathspec==0.11.2
platformdirs==3.10.0
premailer==3.10.0
protobuf==4.23.4
pyannotate==1.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pylint==2.17.5
PyNaCl==1.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.31.0
termcolor==2.3.0
tomli==2.0.1
tomlkit==0.12.1
trio==0.22.2
trio-websocket==0.10.3
typesystem==0.4.1
typing_extensions==4.7.1
undetected-chromedriver==3.5.0
uritemplate==4.1.1
urllib3==1.26.16
websockets==11.0.3
wrapt==1.15.0
wsproto==1.2.0
yaspin==2.4.0
